Separate deal comments from deal_details.dart
